gitlab CI: remove redundant "dependencies" info
When we started using "needs" for the acyclic graph scheduling we
needed to duplicate dependencies and needs (I forgot if it was
documented or just buggy).

Nowadays gitlab doc https://docs.gitlab.com/ee/ci/yaml/#needs says

>In GitLab 12.6 and later, you can’t combine the dependencies keyword
with needs to control artifact downloads in jobs. dependencies is
still valid in jobs that do not use needs.

Some jobs in stage-1 had "dependencies: []" without a corresponding
"needs", not sure why as the only job in a previous stage is
docker-boot which has no artifacts to skip.

We don't want "needs: []" as that would fail to wait for the docker
image, so we just have nothing instead.