[coqide] fix procedure to parse arguments
coqide calls coqidetop -batch, if you are in -async-proof on then
coqidetop spawns a worker and passes -batch to it. At some point,
I could not find the commit, this made the worker die.

On linux it seems it works anyway, but on windows this death is
perceived by coqide which then does not start.