[coqide] fix procedure to parse arguments
coqide calls coqidetop -batch, if you are in -async-proof on then
coqidetop spawns a worker and passes -batch to it. At some point,
I could not find the commit, this made the worker die.

On linux it seems it works anyway, but on windows this death is
perceived by coqide which then does not start.

(cherry picked from commit 065913e830ea7b01a500c523c47dcbcdab980b72)