Remove opam files from repository.
They were not introduced for external use by users but because they
were needed by Dune.  Now that they can be autogenerated by Dune, we
do not need to check them into the repo anymore.  Removing them will
remove the ability for users to rely on them and thus a confusing
experience (due to not having Dune or not getting coqidetop
installed).