Improve type-checking of case predicates
We now verify independently that the match predicate and branches are correct
w.r.t the scrutinees inductive type declaration and effective parameters,
i.e. the predicate is of the form (indices ,, x : Ind@{u} params indices)
for some universe instance u. Only in a second phase we ensure that the
match can be applied to the scrutinee, which just means testing
cumulativity between its type (Ind@{u'}) and the type expected by the
predicate (Ind@{u}). The previous strategy using cumulativity was a tad
too permissive and we couldn't prove completeness for it in MetaCoq in
particular.

Taking a step back, the root of this problem is that we
were taking the discriminee's type as determining the shape of the
predicate/branches types while this type can change during reduction
or by cumulativity. Here we instead take the view that the "eliminator"
is defined independently of its application to the object to eliminate.
Again the change of representation of cases to make the information
currently implicitely present in the predicate and branches would make
this mistake less easy to make (and the implementation simpler as well).

This adapts both the kernel and Pretyping.typing.
Nothing needs to change for Pretyping as the user cannot
specify different params/universes for a match than the ones inferred in
the scrutinee's type. As a result this change should also be entirely
transparent to the user.

Also improve the error message about wrong elimination in Typing.typing,
reusing the same routine as kernel typing.