Fix #13732: Implicit Type vs universes
This returns to the previous behaviour of Implicit Type forgetting
universes.

`Implicit Type T : Type@{u}.` may be confusing as it is equivalent to
`: Type`, but until we have a better idea of what to do with anonymous
types I see no other solution, especially in the short term to fix the
bug.