Do not rely on the template status to decide keeping proof equalities.
Instead, we add a table to selectively activate the Keep Proof Equality option.
This allow decoupling the two feature and easing the eventual removal of
template polymorphism. For backwards compatibility, template inductive types are
implictly added to the table.

A careful analysis of the previous code indicates that this feature is quite
fragile because it depends on the syntactic shape of the type of the argument,
which is not normalized. In particular it means that equality proof preservation
is not stable by conversion and is sensitive to internals of the retyping
algorithm. The new code has the same defect, but we remove the now unused
truncation_style flag hack in order not to tempt any innocent bystander.