[build] Make the dune-based build system the default for developers
Recent experiences plus #13617 have shown that it could be a good idea
to actually default to the dune-based build system for developers.

Workflow is a bit different, but I find it _much_ superior and faster,
at least for all my use cases. YMMV.

Main reasons developers do still need to use the make-based system are:

- documentation? Not sure about this one, but should be easy to fix

- different vo compilation modes, dune doesn't indeed understand
  native, vok, and vio targets. I am not sure if these targets are
  used in development [outside of the test suite, where they are
  properly tested].

  Native will be fixed as soon as we get coqnative, there is also
  support for it in more recent versions of dune, see PR.

  vok support is also planned, see corresponding upstream PR

Closes #7208