[loadpaths] Remove coq-core compat code.
This shouldn't be needed after the switch to Dune for the OCaml parts,
(#13617) as the discrepancy between build systems w.r.t. install paths
have been solved.

This commit just aims to check compatibility with the CI [which will
fail due to the missing overlays for #13617) , we will provide a
better API so all different Coq paths are under the same type in the
upcoming `coq.boot` PR.