Gramlib: new tree-based implementation of finite token look-ahead.
Formerly, token look-ahead was able to consider a linear sequence of
tokens as one token, recovering from a failure as a whole. That is, in
[ "foo"; "bar"; ... -> action | other-rules ], a failure after "foo"
was seen as a failure of parsing the sequence "foo";"bar" and was
going back to other-rules.

This had limits. For instance, parsing "foo";"bar3" in the following
non-linear situation:

[ "foo"; "bar1"; ... -> action
| "foo"; "bar2"; ... -> action
| other-rules ]

was not going back to other-rules.

This commit is about using a tree of tokens rather than a sequence of
tokens, so as to support the above example.