Add better `String_as_OT` and `Ascii_as_OT`
There are versions in `theories/Structures/OrderedTypeEx.v` since
https://github.com/coq/coq/pull/7221, but this module is deprecated, and
there does not seem to be any standard way to convert to the
non-deprecated `OrderedType`.  This fixes that problem, and allows users
to sort strings and ascii.

Pluasibly, we should move these modules to `Ascii.v` and `String.v` (and
fill out `Byte.v`, but this seems to require a deeper restructuring of
the files.  For `N`, `Z`, et. al., there seems to be a clear distinction
between which modules get imported (to get notations, `N`, etc), and
which modules do not get imported (so that lemmas are qualified for what
they are about, so that we have `N.t` and `Z.t` not `t`, etc).  I'm not
sure how to restructure the `Ascii`, `Byte`, and `String` files in this
way, and this reorganization seems beyond the scope of this minor
enhancement.