[printing] Remove ability to set per-sentence "verbosity"
As part of the ongoing effort on
https://github.com/coq/coq/issues/12923 , we simplify the verbose API,
and in particular we remove the ability for `Stm.add` to attach a
verbosity argument per-sentence in favor of the global settings.

Having per-sentence verbosity information to override user-provided
settings seems strange, tho I think this change makes sense, but I
could be missing something.

Note in particular that vernac processing sets verbosity as:
```ocaml
Stm.add ~doc:state.doc ~ontop:state.sid (not !Flags.quiet) com in
```
so it is actually always the value of the global parameter in Coq.

The XML protocol allows to specify it as an argument, as far as I know
CoqIDE uses it but I am not sure the use did make sense in general.