Add `inversion_sigma H` and `inversion_sigma H as ipattern`
Now `inversion_sigma` can be invoked on specific hypotheses, just like
`inversion` and `induction`.

Also fill out the theory of equality of `ex` more

To speed up `inversion_sigma`, we add many special cases so that we
don't have to do as much reduction/unification.

We also remove useless simplification in `inversion_sigma`.

Now that we have specialized induction predicates for the various levels
of constructors, we don't need to simplify away the constructors.