Less crazy implementation of errors in Gramlib.
Instead of printing information on stderr, we display it in the string passed
to the Failure exception. I did not use an algebraic exception on purpose
given that it is unclear whether there is some caller up relying on the exception
being precisely Failure.

This was observed when debugging the Ltac2 grammar anomalies.