[build] [dune] Add world-states target to produce a subset of world
Some users want to have a subset of the world target for faster
interaction with non-dunerized builds.

This is not seamless done in Dune due to install layouts not composing
in general [so implementing `%{pkg:coq:foo}` is hard when the `coq`
package is not in the workspace]

Instead, we provide a convenience target to build some subsets of
world until we improve support in dune.

Likely, Dune should support the `@install` alias scoped to
sub-directories but there seems to be a problem with it as of today.