Statically ensure level creation in Gramlib.
We split the constructor for grammar extension in two, depending on whether
the extension command should create new levels or populate a previously
defined one. In the latter case, level meta-data is ensured not to be
present.

We summarily adapt both coqpp and the grammar extension mechanism to this
change.