[DO NOT MERGE] Commit for testing MacOS Install
I am hopeful that submitting this as a draft PR will help with debugging
and fixing #14228.  If this passes the CI, then I appologize in advance
for the noise and will close #14228.  If this does not pass the CI, then
I hope ejgallego will help me debug what's wrong by inserting the
necessary printouts into this PR.

I've removed the GitLab and Windows CI, since we don't need to test
those to see what's going on with Mac.  I've adjusted only the prefix to
`/usr/local`, which used to work fine before the dune PR without `sudo`
as can be seen at
https://github.com/mit-plv/fiat-crypto/blob/efe1bec26b06b5206fd70fb8164cb757da5c10e5/.github/workflows/coq-macos.yml