[build] Ensure `make world` builds all targets present in `make install`
It is important to ensure a phase separation of `make world` and `make
install` as this last one may be called using `sudo` in some contexts
and thus create build objects under the root user, which leads to
disaster usually.

In order to fix that quickly, we just extend the `world` target to be
comprehensive.

Note that dune upstream enforced this by forbidding `dune install` to
build anything, but there is a bit more convenient as the
`foo.install` file depends on all the objects to installed already by
construction.

We should follow a similar path in our make wrappers, but that would
require a bit of rewrite / design, with given the status of the code
may not be a bad idea as it has gotten to a point where cleanup would
be nice.

Fixes #14228 , keeping #14232 open as to try to improve install
locations a little bit for 8.14, tho the issue is subtle.