Specialize the default position selector semantics in Grammar.
Before this patch, adding rules to an entry without specifying a position
selector would result in two very different behaviours depending on the dynamic
content of the entry.

- If the entry was empty, it would create a fresh level and populate it with the
provided rules.
- If the entry was not empty, it would reuse the topmost level to populate it
with the provided rules. In case the extension also contained a level name or
an associativity status, it would ignore the name and replace the old
associativity by the new one.

This patch introduces a new selector Top to explicit the intent of the
programmer. This selector behaves as in the second bullet above, and fails
if the entry is empty. Dually, we specialize the behaviour of an absent
selector to the first bullet, and fail in case the entry is not empty.