[build] Remove undocumented prefix install vars, use DESTDIR
Coq's makefiles provided an undocumented way to emulate `DESTDIR`,
by means of the `COQINSTALLPREFIX` variable.

This was a hack and only used internally; thus removed recently, but
indeed it turns out that most packaging systems do need to access to
similar functionality and the change broke them.

We thus add support and document the `DESTDIR` variables for Coq's
`make install`, aligning it the standard interface, supported by
coq_makefile and Debian for example.

Fixes #14248