Hide the instantiation of patterns with constr variables under an API.
This makes internal one critical call to the dreaded pattern_of_constr
function. As a nice bonus, this also enforces statically that we never
perform compositions of pattern_of_constr calls. Instantiating a pattern
is just mapping a given set of variables to specific constr that are
turned into constr_patterns, where the instantiation is not recursive.