Slightly improve guard condition in the presence of commutative cuts.
The type-based filtering of arguments passed across a "match" was not
taking into account that the return predicate had a specific
constructor instance in each branch.

We take the pattern-structure of this instance into account (keeping
only constructors).

This allows to recognize more useful fixpoints as guarded (e.g. map3 on
vectors when using small inversion, or #17062).