Fixes #13966: broken invariant when notation not for pat overrides one ok for pat.
This was introduced in 26a456cbf1.

We do not expect a notation overriding another one to have the same
applicability in "match" pattern.

In particular, when a notation interpretation overrides another one,
this deactivates also the use of the previous one in patterns, if ever
the new one does not apply to patterns (and vice-versa).