Prevent a quadratic blowup in Micromega annotations.
The previous code used lists and repeatedly appended them in the wrong order,
leading to O(n²) behaviour, or even worse. Since the result is only used in
the end to build an unordered set, we simply create the annotation trace as
the free datatype closed under concatenation and singleton, i.e. a binary
tree.

The quadratic behaviour was observed in coqprime.