test-suite: don't use $(wildcard) in approve-output
It is evaluated before running the dependencies, so from a clean state
it evaluates to nothing even if some tests end up failing and
approve-output has to be run twice.

Instead we are robust to not needing to work (with the glob being
evaluated to itself) by adding a "if [ -f $f ]" evaluated by the
shell.