Deprecate the new auto tactic.
It was not documented and not used in the CI either. It seems to have
been introduced by e961ace3 but never really took off. Since this is
responsible for a big chunk of code in Auto, it looks like it is worthwhile
to deprecate it in order to remove it as soon as possible.