[coqnative] Support -boot option
`coqnative` must take `-boot` in order to work without the stdlib
present, this is required when bootstrapping Coq, or for HoTT for
example.

We also take -noinit as to provide a more uniform interface with the
rest of the tools, waiting for the consolidation of this code in a
common lib.