Slightly saner definition of Unification.subst_defined_metas_evars.
It seems that the intent is to fail whenever the term contains a meta,
but the evar case introduced by c9728f6 was catching the error probably
incorrectly. The commit message indicates that the goal of this patch was
to catch the error when the evar is absent, not when the expanded term
contains a meta.