coq_makefile: Handle single quotes in `-arg` itself
Previously, `make` and the shell did process the arguments passed using
`-arg` further. For example parsing escape sequences or expanding
variables. By properly escaping `COQMF_OTHERFLAGS`, this behaviour is
avoided and `coqProject_file.ml` is now in control of the arguments.
Coqc now receives exactly the list of strings in `proj.extra_args`.

The handling of single quotes `'` is retained. This was a way with which
arguments containing spaces could be passed to coqc via a `_CoqProject`
file. The behaviour of the shell on these strings was replicated, and
users of `coq_makefile` shouldn't notice a difference.
Other users of `coqProject_file.ml` like CoqIDE now experience `-arg`
the same way as `coq_makefile` does.

The function `makefile_quote` is not suitable, if the quoted string
should ever be the name of a Makefile-target, because then other rules
apply which would break the Makefile, if the passed string contains
spaces or `%`.