Fixes #4167: know that main argument of projection is explicit in t.(f) notation.
We force the main argument of projection to be explicit in the t.(f)
notation and treat separately the implicit parameters and the implicit
extra arguments (if any). In particular, we check statically that "f"
has the expected number of explicit parameters.

In addition to fixing #4167, it improves the error message when fields
do not take right number of explicit args. Previously this was
detected only at typing time resulting in unclear messages saying that
the argument "x" of the projection in syntax "x.(f)" had a wrong type.