[build] [test-suite] Move `fake_ide` to test-suite
Being only used for testing and nothing else, the `fake-ide` binary can
go live in the test-suite/ directory.

Behavioral change: `fake-ide` now needs an additional cmd-line argument:
the path to `coqidetop`. Previously both binaries lived in the same
folder, so this wasn't a problem. Now this information has to be passed
through by the Makefiled.

Closes #14306
Co-authored-by: Emilio Jesús Gallego Arias <e+git@x80.org>