Do not apply discriminated unfold hints on arbitrary goals.
The current behaviour was to indiscriminately try to apply on any goal
unfold hints for transparent constants from discriminated databases, since
they were not indexed by the head constant. In particular, this might have
unfolded the constant even if it was at the head of the goal, contrarily to
what the documentation claimed.

I think it was written this way out of fear of the dnet pattern algorithm. In
discriminated mode, transparent constants are considered to match everything,
but maybe the author of the code was confused by the complexity of the dnet
implementation.

This patch changes the semantics of auto tactics in fringe cases, and should
be mostly backwards compatible.

Fixes #14874: Exponential blowup with Hint Unfold.