Added documentation for `#[projections(primitive)]` attribute.
We add documentation to the refman mentioning the new
`#[projections(primitive)]` attribute. We also do some small
refactoring in the primitiive projections chapter. The mention of
"primitive records" has been removed since I find this confusing
terminilogy. It has nothing to do with the "primitive" datatypes found
elsewhere and we are really only talking about the projections when
mentioning this feature anyhow.

Replaced "macro" with :cmd:`command`.

Co-authored-by: Jim Fehrle <jim.fehrle@gmail.com>