More uniform handling of variable patterns in Btermdn.
We first check that a pattern variable has a body before looking it up in the
transparent state. This only affects discriminated nets. The behaviour or
non-discriminated nets is very weird and asymmetrical since in patterns variables
match everything but in terms they match only variables, regardless of the existence
of a body.