Remove the eta flag from CClosure redflags.
This flag was never used in the kernel, and the only user was the shrink_eta
function which did not use it with any other reduction flag. This caused quite
a bit of duplication in the various upper layer machines for nothing.
Furthermore the semantics was already dubious in some corner cases with the eta
flag alone, but it would have been worse if trying to mix eta with additional
reductions.

We inline the only call in the shrink_eta function, and take advantage of the
operation to make it correct w.r.t. evars.