On-the-fly η-reduction of dnet patterns and terms.
Due to the way opaque terms are handled, dnet pattern matching was failing
to recognize that the η-expansion of a term was matching that term, in both
directions. To solve this, we simply η-reduce both terms and patterns before
adding and retrieving them from the dnet.

Fixes #14731: Term matching in discriminated hints is not up-to η-expansion.