Split Scheme Equality into own vernac + misc cleanup
The generation of equality schemes and induction schemes seem to have
both been bundled together in order to share the vernac command. This
however causes a lot of unnecessary work down the line.

We therefore separate these two (very separate) processes. We add a
`Scheme Equality for` command which allows us to use the same vernac
command as before, but now importantly it doesn't share any work with
the induction schemes. The code of the induction schemes has been
simplified and cleared up.

There used to be a function called `split_schemes` which was
separating equality schemes from induction schemes and extracting some
data from the induction schemes. This has now been simplified and
renamed to `name_and_process_schemes` and the useless work has been
removed.

Separating this data does propagate throughout the code but I have
opted to not be too destructive, especially with functions about the
actual scheme generation.