Add new, more permissive criterion for instantiating self-referencing evars.
For an evar instance ?x := C[?x] referring to itself being valid, it
is enough that C[C[?x]] is convertible to C[?x]. For that it is enough
that all subterms of C where ?x occurs are "local" eta-expansion of ?x.

This includes e.g.:
- C[] = []
- C[] = λx.([] x)
- C[] = S (pred [])
- C[] = {fst := fst []; snd := 0}

This commit implements a test checking when the criterion holds.

Pending questions: would it be worth to eta-reduce the solution to
reduce its size, when applicable?