Tweak the heuristic for frozen evars in Equality subterm matching.
As explained by a comment, we used to consider all evars appearing in the
type of the clenv to be frozen as an overapproximation of evars freshly
introduced by lemma instantiation. For subterm matching though, we know
statically that no fresh evar can be introduced because the clenv was
generated just above.