Correctly set prefix when no default is given.
The behavior before was defaulting on the deprecated "local"
option. This doesn't make sense since "-local" is just an alias for
"-profile devel" which sets the prefix anyway. Therefore, when no
prefix is given, as is often the case, we should use the default
prefix which can be found in `do_one_instdir`. Of course, this is only
a hack to get things working for the time being. The correct solution
would be to rewrite the entire configure.ml file to not have such
ad-hoc reasoning.