Stop relying on undiscriminated ground checks in Btermdn.
This removes a very weird part of the code which was probably there for
bad reasons. Notably, the dnet matching algorithm was so ad-hoc that it
would match (or not) many things erroneously, so that the ground check
was there allow to bypass those bugs without much fuss.

This seems to be more efficient as showed by the bench.