More correct implementation of dnet matching for case nodes.
We try to move closer to an overapproximation of neutral forms for
βι-reduction, up to η-conversion. Before, case nodes were incorrectly
mapped to Nothing, even if they could have ι-reduced to an arbitrary
term. Now we consider them to match anything, which is a bit violent
and should probably be handled in a more subtle manner in a future
patch.

We also handle undiscriminated β-redexes as arbitrary patterns for good
measure.

Fixes #13209: TC search treats match (if) as TC-opaque.