Make explicit that unfold_reference only depends on the transparent state.
There was a confusion between the reduction flags and the transparent state.
It turns out we only need the latter to decide whether to unfold a
reference, since when *converting* RelKey we *must not* check the delta flag,
i.e. we unfold unconditionally.

This highlights another suspicious point, which is that fDELTA has little
to do with what is usually called δ-reduction, i.e. constant unfolding. This
flag only controls unfolding of rel variables from the external environment.
It does not even affect unfolding of named variables.