Correctly handle delta flags for CClosure reduction.
There seems to be a pervasive confusion around the ill-named delta flag.
Some places use it with the clear intent to mean that it should allow
unfolding of constants in the kernel, which is not what it was doing.
Thankfully these calls were often wrapped in modifications of the
transparent state through CClosure.RedFlags.red_add_transparent, which
worked around the borken semantics of the delta flag.

Some dubious calls to the redflags had to be fixed.