[configure] Detect Dune 2.9, use right install options in this case.
We check Dune's version with configure, and provide a warning if Dune
< 2.9 about it ignoring `-docdir` and `-configdir` options.

We also mention the easiest workaround, using `opam-install` which
supports these options correctly.

Note that as Coq now consists of multiple packages, the `-docdir`
option has been changed to denote the root for the documentation of
all packages, as opposed to the previous behavior where it was used to
point to the absolute documentation path.

Should help with #14232 and #14833