Don't bench opam's overhead.
We do the perf and time inside the "wrap-build-commands" hook system
of opam, instead of wrapping the whole opam call.

A few env variables are used to communicate with the wrapper, see wrapper.sh

A little care is needed to handle packages with multiple build
commands (eg corn does configure; make): we sum the command timings in
render_results.

Tested at https://gitlab.com/SkySkimmer/coq/-/jobs/1568571689 and
https://gitlab.com/SkySkimmer/coq/-/jobs/1568718247 on the new
machines.
The new machine instability does not appear, which is the main goal of
this PR.
Package timings are less than previously so small instabilities can
produce bigger percentages, eg in the 2nd run ssreflect reports -1.48%
but this is just 0.4s.
(ignore the "coq" job, it is not printed in normal operation and is
just there to debug multi-command packages)

Perhaps we should also be wrapping the install commands, however it's
unclear how much information that would give.