[build] Move core OCaml libraries under `core`
As discussed in the Coq Call, this is a step towards #6137 .

This is the most painful step to do, but still that issue is not
closed, as we can hope for a few other files at root to go away, such
as the makefiles, the .opam files, `install.sh`, the `user-contrib`
directory, and likely `coqpp`, which could be placed under `tools`.

We may also move `plugins` once #14059 lands in main.