Split the universes_entry type in two variants.
The first one is kernel-facing, and has an empty field for the monomorphic
case. This statically ensures that it is not possible to add global
constraints by adding a constant.

The second one is for the upper levels, and allows an implicit declaration
of monomorphic constraints. This makes the API more palatable.

For now the binding mechanism of monomorphic universes is broken, leading
to two failures in output tests.