Officially remove the printing of global universes for constants.
This was misleading, since the printing suggested that the universes
were quantified upon, which was not the case. Furthermore, relying on
the Monomorphic field was incorrect for side-effects, whose constraints
were not generated the same way.

We could reintroduce it by relying instead on DeclareUniv quite easily, but
I think this would still be counterproductive. We already have access to the
fancy printing based on the name of the constant that declared it which looks
enough for most practical purposes.