[build] Set dune root explicitly / use --release
This should fix #14915 , the problem arises with layouts of the form:

```
foo/dune-project
foo/_opam/bar/dune-project
```

so when inside `foo/_opam/bar`, dune will switch to `foo` as the
project root, for example for `dune build` or `dune exec` , however,
as `_opam` is an ignored dir by default in the `foo` context, the
command will fail.

So indeed, it turns out `--profile release` is not enough, but
`--release` must be used when configured for install.

Thanks a lot to Kate for the help!