Correctly cache opaque body access in Print Assumptions.
The previous state of the code forced a recomputation of a huge proofterm
every time a constant was accessed, even if the body was immediately discarded.
We now correctly check for presence in the cache before forcing the
recomputation.

(cherry picked from commit bb7524cf32fa3183a5d6da2d3ed9049d99633050)